[Update] index_v2.html
For my version 2 of the index (home page) I added in a navigation, with different links to each page that I will make. I also added a header, and displayed the text, 'Audio Interface.' And I added a second header displaying the text, "Callan Potgieter." I also brushed up some of the other code to make it look neater.